add comentCount and deal with some Exception

+++ src/main/java/life/cris/community/controller/QuestionController.java
@@ -1,6 +1,7 @@
 package life.cris.community.controller;
 
 import life.cris.community.dto.QuestionDTO;
+import life.cris.community.mapper.QuestionExtMapper;
 import life.cris.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -14,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name="id") Integer id,
+    public String question(@PathVariable(name="id") Long id,
                            Model model){
         QuestionDTO questionDTO= questionService.getById(id);
         //累加评论数
